Reorganize Dockerfile to be more efficient
When your Dockerfile has `COPY . ./` instruction, a change to any file
in your directory will invalidate the cache from that point forward. That
means you want to accomplish as much as possible *before* the `COPY`
command.

This commit rearranges the Dockerfile to perform all apt and pip installs
earlier in the file to make local rebuilds go much faster.